+++ TabletopTweaks-Base/NewContent/MythicFeats/TitanStrike.cs
@@ -1,4 +1,5 @@
-﻿using Kingmaker.Blueprints.Classes;
+﻿using Kingmaker.Blueprints;
+using Kingmaker.Blueprints.Classes;
 using TabletopTweaks.Core.NewComponents.AbilitySpecific;
 using TabletopTweaks.Core.Utilities;
 using static TabletopTweaks.Base.Main;
@@ -6,6 +7,10 @@ using static TabletopTweaks.Base.Main;
 namespace TabletopTweaks.Base.NewContent.MythicFeats {
     static class TitanStrike {
         public static void AddTitanStrike() {
+            var StunningFistOwnerBuff = BlueprintTools.GetBlueprintReference<BlueprintBuffReference>("d9eaeba5690a7704da8bbf626456a50e");
+            var StunningFistOwnerFatigueBuff = BlueprintTools.GetBlueprintReference<BlueprintBuffReference>("696b29374599d4141be64e46a91bd09b");
+            var StunningFistOwnerSickenedBuff = BlueprintTools.GetBlueprintReference<BlueprintBuffReference>("4d7da6df5cb3b3940a9d96311a2dc311");
+
             var ImprovedUnarmedStrikeMythicFeat = BlueprintTools.GetBlueprint<BlueprintFeature>("e086a07dae105244291fb11e05d0715f");
             var TitanStrikeFeature = Helpers.CreateBlueprint<BlueprintFeature>(TTTContext, "TitanStrikeFeature", bp => {
                 bp.m_Icon = ImprovedUnarmedStrikeMythicFeat.m_Icon;
@@ -17,7 +22,13 @@ namespace TabletopTweaks.Base.NewContent.MythicFeats {
                 bp.IsClassFeature = true;
                 bp.Ranks = 1;
                 bp.Groups = new FeatureGroup[] { FeatureGroup.MythicFeat };
-                bp.AddComponent<TitanStrikeComponent>();
+                bp.AddComponent<TitanStrikeComponent>(c => {
+                    c.m_StunningFistBuffs = new BlueprintBuffReference[] {
+                        StunningFistOwnerBuff,
+                        StunningFistOwnerSickenedBuff,
+                        StunningFistOwnerFatigueBuff
+                    };
+                });
                 bp.AddPrerequisiteFeature(ImprovedUnarmedStrikeMythicFeat);
             });
             if (TTTContext.AddedContent.MythicFeats.IsDisabled("TitanStrike")) { return; }