+++ TabletopTweaks-Base/UMMSettingsUI.cs
@@ -93,6 +93,10 @@ namespace TabletopTweaks.Base {
                     ("Base", Fixes.Shaman.Base),
                     Fixes.Shaman.Archetypes
                 );
+                SetttingUI.NestedSettingGroup("Skald", TabLevel, Fixes.Skald,
+                    ("Base", Fixes.Skald.Base),
+                    Fixes.Skald.Archetypes
+                );
                 SetttingUI.NestedSettingGroup("Slayer", TabLevel, Fixes.Slayer,
                     ("Base", Fixes.Slayer.Base),
                     Fixes.Slayer.Archetypes